create reset_database
