+++ backend/database.go
@@ -24,130 +24,6 @@ func (d *Database) Close() error {
 	return d.db.Close()
 }
 
-// SetupTables creates necessary tables if they do not exist.
-func (d *Database) SetupTables() error {
-	// TODO maybe we should change all ID fields to string type for readability
-	commands := [...]string{
-		`
-		CREATE TABLE IF NOT EXISTS Vendor (
-			ID BINARY(16) NOT NULL,
-			Name VARCHAR(100) NOT NULL,
-			BusinessAddress VARCHAR(500) NULL,
-			Website VARCHAR(500) NULL,
-			BusinessHours VARCHAR(500) NOT NULL,
-			Phone VARCHAR(50) NULL,
-			BusinessLogo BINARY(16) NOT NULL,
-			Latitude FLOAT NOT NULL,
-			Longitude FLOAT NOT NULL,
-			PRIMARY KEY (ID)
-		)
-		`,
-		`
-		CREATE TABLE IF NOT EXISTS User (
-			ID BINARY(16) NOT NULL,
-			Email VARCHAR(100) NULL,
-			Username VARCHAR(100) UNIQUE NULL,
-			FirstName VARCHAR(100) NULL,
-			LastName VARCHAR(100) NULL,
-			SignUpDate DATETIME NULL,
-			LoginPassword BINARY(32) NULL,
-			UserType TINYINT NULL,
-			Photo BINARY(16),
-			PRIMARY KEY (ID)
-		)
-		`,
-		`
-		CREATE TABLE IF NOT EXISTS Reviews (
-			ID BINARY(16) NOT NULL,
-			Text text(500) NULL,
-			VendorID BINARY(16) NOT NULL,
-			UserID BINARY(16) NOT NULL,
-			DatePosted DATETIME NULL,
-			PRIMARY KEY (ID),
-			FOREIGN KEY (VendorID) REFERENCES Vendor(ID)
-			ON DELETE CASCADE ON UPDATE CASCADE,
-			FOREIGN KEY (UserID) REFERENCES User(ID)
-			ON DELETE CASCADE ON UPDATE CASCADE
-		)
-		`,
-	}
-
-	for _, command := range commands {
-		_, err := d.db.Exec(command)
-		if err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-// AddTestData adds test data. This is only for testing purposes.
-func (d *Database) AddTestData() error {
-	// Check if Vendors data exists
-	vendorCount := 0
-	if err := d.db.QueryRowx("SELECT COUNT(*) FROM Vendor;").Scan(&vendorCount); err != nil {
-		return err
-	}
-
-	if vendorCount < 2 {
-		vendor0 := &Vendor{
-			ID:              uuid.MustParse("e72ac985-3d7e-47eb-9f0c-f8e52621a708"),
-			Name:            "vendor0",
-			BusinessAddress: "address0",
-			Website:         "www.vendor0.com",
-			BusinessHours:   "Mon-Sun 8:00AM-5:00PM",
-			Phone:           "123-123-1234",
-			BusinessLogo:    "image0_url",
-			Latitude:        "47.608013",
-			Longitude:       "-122.335167",
-		}
-
-		if err := d.VendorCreate(vendor0); err != nil {
-			return err
-		}
-
-		vendor1 := &Vendor{
-			ID:              uuid.MustParse("b924349d-442f-4fff-984e-ab0ec36f4590"),
-			Name:            "vendor1",
-			BusinessAddress: "address1",
-			Website:         "www.vendor1.com",
-			BusinessHours:   "Mon-Sun 8:00AM-5:00PM",
-			Phone:           "321-321-4321",
-			BusinessLogo:    "Image1_url",
-			Latitude:        "47.982567",
-			Longitude:       "-122.193375",
-		}
-
-		if err := d.VendorCreate(vendor1); err != nil {
-			return err
-		}
-	}
-
-	userCount := 0
-	if err := d.db.QueryRowx("SELECT COUNT(*) FROM User;").Scan(&userCount); err != nil {
-		return err
-	}
-
-	if userCount < 1 {
-		user0 := &User{
-			ID:         uuid.MustParse("02c353e2-e0f5-4730-89c7-b0a0610232e4"),
-			Email:      "test@example.com",
-			Username:   "test",
-			FirstName:  "Selina",
-			LastName:   "Tan",
-			SignUpDate: time.Now(),
-			UserType:   0,
-			Photo:      uuid.MustParse("3959ac2f-756e-4632-9678-912130384248"),
-		}
-
-		if err := d.UserCreate(user0, "password"); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
 type Vendor struct {
 	ID              uuid.UUID
 	Name            string
