+++ backend/main.go
@@ -10,13 +10,17 @@ import (
 )
 
 func main() {
-	database := getDatabase()
-	if err := database.SetupTables(); err != nil {
+	config := mysql.Config{
+		User:                 "root",
+		AllowNativePasswords: true,
+		DBName:               "streetfoodlove",
+		ParseTime:            true,
+	}
+	db, err := sqlx.Connect("mysql", config.FormatDSN())
+	if err != nil {
 		panic(err)
 	}
-	if err := database.AddTestData(); err != nil {
-		panic(err)
-	}
+	database := NewDatabase(db)
 
 	api := API{&Backend{database}}
 	defer api.Close()
@@ -36,46 +40,4 @@ func main() {
 	if err := server.ListenAndServe(); err != nil {
 		panic(err)
 	}
-}
-
-// getDatabase creates database and returns the database connection.
-func getDatabase() *Database {
-	config := mysql.Config{
-		User:                 "root",
-		AllowNativePasswords: true,
-	}
-	db, err := sqlx.Connect("mysql", config.FormatDSN())
-	if err != nil {
-		panic(err)
-	}
-
-	commands := [...]string{
-		`
-		CREATE DATABASE IF NOT EXISTS streetfoodlove
-		CHARACTER SET utf8mb4
-		COLLATE utf8mb4_unicode_ci;
-		`,
-		`USE streetfoodlove;`,
-	}
-
-	for _, command := range commands {
-		_, err := db.Exec(command)
-		if err != nil {
-			panic(err)
-		}
-	}
-
-	db.Close()
-
-	config = mysql.Config{
-		User:                 "root",
-		AllowNativePasswords: true,
-		DBName:               "streetfoodlove",
-		ParseTime:            true,
-	}
-	db, err = sqlx.Connect("mysql", config.FormatDSN())
-	if err != nil {
-		panic(err)
-	}
-	return NewDatabase(db)
 }
